Added crowdsource data from https://docs.google.com/spreadsheets/d/1RXMhqzOZDqygWzyE4HXi9DnJnxjdp0NOhlHcB5SrSZo/edit?usp=sharing
Preliminary commit to add crowdsourced data to the old version of the repo before doing a re-organisation. This adds the public version of crowdsourced data cleanup as described at http://access.okfn.org/2014/04/01/wellcome-trust-apc-data-thank-you/ to the repo. This commit is included so as to maintain a version of the overall history of data sources used for this project.